#3jdhg8 preparing for PUT method to update mongodb entry for a product
